etl/requirements.txt
pandas==2.1.4
pdfplumber==0.10.3
camelot-py[cv]==0.10.1
tabula-py==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
PyYAML==6.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
boto3==1.34.0
playwright==1.46.0
